Introduce concept of username
Prior to this commit application was handling different clients
on the WebSocket level. Meaning each WebSocket object was unique
therefore no need of introducing synthetic clients differentiator.

Username must have been added to application since the pairing
feature requires to identify clients by the username. Connected
player MUST be distinguishable from each other.

Important Note
Current implementation is NOT and should NOT be treated as a way
of authentication and authorization to the server. This implementation
is only to provide pairing feature. Server is STILL not providing
a secure and reliable way of authentication and authorization as
server is still under active development.