Refactor code under domain package
Before the commit the code was unnecessary handling creation of
url's in the RedirectEndpoint class. This responsibility is easy to
meet in the caller code. Before the commit the code was
unnecessarily storing ConnectedClient reference in GameBoardState.

After the commit the code is cleaner by removing redundant files.
Also the code is much more cohesive since it does not reference types
defined in different modules.